Object  Recognition through the numbers of Corners
This is a experimental software. It shows how hough can be used for this purpose. it has it limits when the number of  corners increase. It hits the limit of what hough can handle also at this point the possibility to use angles for recognition isn't implemented yet. (In this case it could only do concave objects since there is no order for the edges and corners)